fix: bkmr integration

+++ plugin/vimania_uri_rs.vim
@@ -88,13 +88,17 @@ command! -nargs=0 VimaniaThrowError call VimaniaThrowError()
 
 function! VimaniaDeleteTwbm(args)
   call TwDebug(printf("Vimania args: %s", a:args))
-  " Prompt the user for confirmation
-  let l:confirmation = input("Delete also from bkmr database? (y/n): ")
-  if l:confirmation ==? 'y'
-    python3 xUriMgr.delete_twbm(vim.eval('a:args'))
-    echo "Bookmark deleted."
-  else
-    echo "Deletion canceled."
+
+  " Check if vimania URI twbm integration is enabled
+  if exists('g:vimania_uri_twbm_integration') && g:vimania_uri_twbm_integration == 1
+    " Prompt the user for confirmation
+    let l:confirmation = input("Delete also from bkmr database? (y/n): ")
+    if l:confirmation ==? 'y'
+      python3 xUriMgr.delete_twbm(vim.eval('a:args'))
+      echo "Bookmark deleted."
+    else
+      echo "Bookmark not deleted from bkmr."
+    endif
   endif
 endfunction
 command! -nargs=1 VimaniaDeleteTwbm call VimaniaDeleteTwbm(<f-args>)
